Permitir decimales, en peso total de la evaluacion. en solicitud de cambio de sistema calificacion.

+++ src/main/webapp/public/app/academico/docente/cargaacademica/nuevoSistema.js
@@ -60,8 +60,8 @@ $(function () {
             var pesoEval = $("[name='evaluacionPlan[" + i + "].pesoEvaluacion']");
 
             if ($.isNumeric(pesoTotal.val()) && $.isNumeric(cantEvals.val())) {
-                var pesoTotalNumber = parseInt(pesoTotal.val())
-                var cantEvalsNumber = parseInt(cantEvals.val());
+                var pesoTotalNumber = parseFloat(pesoTotal.val())
+                var cantEvalsNumber = parseFloat(cantEvals.val());
 
                 if (anularNotMin.prop('checked')) {
                     cantEvalsNumber--;
@@ -72,7 +72,7 @@ $(function () {
                  return;
                  }
                  */
-                var pesoEvalsNumber = parseInt(pesoTotalNumber) / parseInt(cantEvalsNumber);
+                var pesoEvalsNumber = parseFloat(pesoTotalNumber) / parseFloat(cantEvalsNumber);
                 pesoEval.val(pesoEvalsNumber.toFixed(2));
             }
         },
